warning fixes

+++ src/pf_imgui/elements/SimpleCurveEditor.cpp
@@ -20,11 +20,11 @@ void SimpleCurveEditor::setMaxPointCount(std::size_t count) { curveData.resize(c
 std::size_t SimpleCurveEditor::getMaxPointCount() const { return curveData.size() - 1; }
 
 float SimpleCurveEditor::getCurveValue(float x) const {
-  return ImGui::CurveValue(x, curveData.size() - 1, curveData.data());
+  return ImGui::CurveValue(x, static_cast<int>(curveData.size() - 1), curveData.data());
 }
 
 float SimpleCurveEditor::getSmoothCurveValue(float x) const {
-  return ImGui::CurveValueSmooth(x, curveData.size() - 1, curveData.data());
+  return ImGui::CurveValueSmooth(x, static_cast<int>(curveData.size() - 1), curveData.data());
 }
 
 void SimpleCurveEditor::renderImpl() {
